Added the changes , login test case
