+++ .gitignore
@@ -2,3 +2,4 @@
 .env
 /dist
 /.scannerwork
+/coverage
