+++ src/modules/organization/input.ts
@@ -109,3 +109,27 @@ export class OrganizationIdInput{
   @Field()
   id!:string
 }
+@InputType()
+export class UpdateOrganizationStatusInput{
+    @Field()
+    id!:string
+
+    @Field()
+    status!:string
+}
+@InputType()
+export class UpdateOrganizationPasswordInput{
+    @Field()
+    id!:string
+
+    @Field()
+    oldPassword!:string
+
+    @Field()
+    newPassword!:string
+}
+@InputType()
+export class DeleteOrganizationInput{
+    @Field()
+    id!:string
+}
