+++ src/modules/organization/organization.response.ts
@@ -147,4 +147,74 @@ export class UpdatJobAppliedStatusResponse{
     @Field()
     status!:string
   }
-  
+  @ObjectType()
+export class UpdateOrganizationStatusResponse{
+    @Field()
+    id!:string
+
+    @Field()
+    status!:string
+}
+@ObjectType()
+export class UpdateOrganizationPasswordResponse{
+    @Field()
+    update_password_state!:boolean
+}
+@ObjectType()
+export class DeleteOrganizationResponse{
+    @Field()
+    id!:string
+}
+@ObjectType()
+class AdminOganizationUser{
+    @Field()
+    id!:string
+
+    @Field()
+    name!:string
+
+    @Field()
+    email!:string
+
+    @Field()
+    phone!:string
+
+    @Field()
+    role!: string;
+}
+@ObjectType()
+export class AllApprovedOrganization{
+    @Field()
+    id!:string;
+
+    @Field()
+    website!:string;
+
+    @Field()
+    description!: string;
+
+    @Field()
+    status!:string;
+
+    @Field()
+    location!:string;
+
+    @Field()
+    created_at!: Date;
+
+    @Field({ nullable: true })
+    updated_at?: Date;
+
+    @Field({ nullable: true })
+    deleted_at?: Date;
+
+    @Field()
+    organization_id!: string;
+
+    @Field()
+    update_password_state!: boolean;
+
+    @Field(() => AdminOganizationUser)
+    user!: AdminOganizationUser;
+
+}
