--- src/modules/s3/s3.input.ts
@@ -1,17 +1,0 @@
-import { Field, InputType } from "type-graphql";
-
-@InputType()
-export class UploadPdfInput{
-    @Field()
-    fileName!:string
-
-    @Field()
-    fileType!:string
-}
-@InputType()
-export class DownloadPdfInput{
-    @Field()
-    key!:string
-
-    
-}
