+++ src/modules/user/user.resolver.ts
@@ -1,10 +1,12 @@
-import { Arg, Mutation, Query, Resolver , Authorized } from "type-graphql";
+import { Arg, Mutation, Query, Resolver , Authorized, ID } from "type-graphql";
 import { User } from "./entity/user.entity";
 import { UserService } from "./user.service";
 import {
+  DownloadPdfInput,
   JobAppliedByUserInput,
   JobApplyInput,
   LoginInput,
+  UpdateJobPostStatusInput,
   UpdateUserInput,
   UploadResumeInput,
   UserIdInput,
@@ -14,16 +16,21 @@ import {
 import { getRepository } from "typeorm";
 import { Service } from "typedi";
 import {
+  DeleteUserResponse,
+  DownloadPdfResponse,
+  GetAllUser,
   JobAppliedByUserResponse,
   JobApplyResponse,
   JobPostResponse,
   LoginResponse,
+  UpdateJobPostStatusResponse,
   UploadResumeResponse,
   UserDetailsResponse,
   WithdrawApplicationResponse,
 } from "./response";
-import { UploadPdfResponse } from "modules/s3/s3.response";
-import { UploadPdfInput } from "modules/s3/s3.input";
+import { UploadPdfResponse } from "../user/response";
+import { UploadPdfInput } from "../user/input";
+import { downloadPdf, uploadPdf } from "../s3/s3service";
 console.log("the resolver of user");
 
 @Resolver()
@@ -90,8 +97,47 @@ export class UserResolver {
   {
     return this.userService.uploadResume(input);
   }
- 
+  @Mutation(() => UpdateJobPostStatusResponse)
+  async updateJobPostStatus(@Arg('input') input: UpdateJobPostStatusInput) {
+    return this.userService.updateStatus(input);
+  }
+  @Query(() => Number)
+  async countUsers():Promise<Number>
+  {
+      return this.userService.countUsers();
+  }
+  @Mutation(() => DeleteUserResponse)
+  async deleteUser(@Arg("id", () => ID) id:string) : Promise<DeleteUserResponse>
+  {
+      return this.userService.deleteUser(id);
+  }
+  @Query(() => [GetAllUser])
+  async users() : Promise<GetAllUser[]>{
+    return this.userService.users();
+  }
+  @Mutation(() => UploadPdfResponse)
+  async generateUploadUrl(
+    @Arg("input") input: UploadPdfInput
+  ): Promise<UploadPdfResponse> {
+    const presignedUrl = await uploadPdf(
+      process.env.AWS_S3_BUCKET_NAME as string,
+      input.fileName
+    );
+    return {
+      presignedUrl,
+      key: input.fileName,
+      publicUrl: `https://${process.env.AWS_S3_BUCKET_NAME}.s3.${process.env.AWS_REGION}.amazonaws.com/${input.fileName}`,
+    };
+  }
 
-  
+  @Query(() => DownloadPdfResponse)
+  async generateDownloadUrl(@Arg("input")input:DownloadPdfInput):Promise<DownloadPdfResponse>
+  {
+    const downloadUrl = await downloadPdf(process.env.AWS_S3_BUCKET_NAME as string ,input.key)
+    return { 
+        downloadUrl,
+        expiresAt: new Date(Date.now() + 3600 * 1000)
+    }
+  }
 
 }
